HimalayaStore/owner/models.py
from django.db import models
from django.db import models
from django.conf import settings
from rest_framework.authtoken.models import Token
import datetime
from django.utils.timezone import now,localtime
from django.core.validators import FileExtensionValidator
from django.db.models.signals import post_save
from django.dispatch import receiver
# Create your models here.
