HimalayaStore/requirements.txt
asgiref==3.2.10
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
cryptography==3.0
Django==3.0.5
django-cors-headers==3.4.0
django-filter==2.3.0
djangorestframework==3.11.0
djongo==1.3.3
dnspython==1.16.0
html5lib==1.1
idna==2.10
Pillow==7.2.0
pycparser==2.20
PyJWT==1.7.1
pymongo==3.10.1
PyPDF2==1.26.0
pytz==2020.1
reportlab==3.5.47
requests==2.24.0
schedule==0.6.0
six==1.15.0
sqlparse==0.2.4
urllib3==1.25.10
webencodings==0.5.1
xhtml2pdf==0.2.4
